+++ java/src/simoil/estrategias/condicionDeFin/EstrategiaCondicionDeFin.java
@@ -6,6 +6,6 @@ import simoil.EmprendimientoPetrolifero;
 public abstract class EstrategiaCondicionDeFin {
 
     public abstract boolean hayQueFinalizar(EmprendimientoPetrolifero emprendimientoPetrolifero,
-                                            int diasTranscurridos, int maximoDias, float porcentajeCriticoPetroleo);
+                                            int diasTranscurridos, int maximoDias, double porcentajeCriticoPetroleo);
 
 }
