+++ java/src/simoil/estrategias/condicionDeFin/EstrategiaCondicionDeFinPorDilucionCritica.java
@@ -5,7 +5,7 @@ import simoil.EmprendimientoPetrolifero;
 public class EstrategiaCondicionDeFinPorDilucionCritica extends EstrategiaCondicionDeFin {
 
     @Override
-    public boolean hayQueFinalizar(EmprendimientoPetrolifero emprendimientoPetrolifero, int diasTranscurridos, int maximoDias, float porcentajeCriticoPetroleo) {
+    public boolean hayQueFinalizar(EmprendimientoPetrolifero emprendimientoPetrolifero, int diasTranscurridos, int maximoDias, double porcentajeCriticoPetroleo) {
         return emprendimientoPetrolifero.yacimiento().composicionDeProducto().porcentajePetroleo() <= porcentajeCriticoPetroleo
                 || diasTranscurridos >= maximoDias;
     }
