+++ java/src/simoil/estrategias/condicionDeFin/EstrategiaCondicionDeFinPorMaximoDias.java
@@ -5,7 +5,7 @@ import simoil.EmprendimientoPetrolifero;
 public class EstrategiaCondicionDeFinPorMaximoDias extends EstrategiaCondicionDeFin {
 
     @Override
-    public boolean hayQueFinalizar(EmprendimientoPetrolifero emprendimientoPetrolifero, int diasTranscurridos, int maximoDias, float porcentajeCriticoPetroleo) {
+    public boolean hayQueFinalizar(EmprendimientoPetrolifero emprendimientoPetrolifero, int diasTranscurridos, int maximoDias, double porcentajeCriticoPetroleo) {
         return diasTranscurridos >= maximoDias;
     }
